Waybar and small changes
Moved around where the fonts were located, they are a desktop thing and
should be there instead of in the general settings.

Added waybar and started configuring it.
I need to figure out how to make nix restart it properly, as it keeps
failing and I have to do it manually.