Some changes, changed from Make to Just, hardware-configuration changes.
First, I changed the Makefile to a justfile because I was not really
using any make features, and it will be easier for further use.

I removed the device specific hardware configuration in favor for a
global one. This should not be uploaded to git but generated as needed
as anytime you would need it, you are redoing the whole device, so just
re-create it.

That should be about it.