Fixed hm build fail and small changes
home-manager was failing to build firefox because firefox would replace
the search engine config file. There is a option in hm
`programs.firefox.profiles.<name>.search.force`
that is used to fix this so that hm won't fail. That just means I have
to manually add search engines. (well the plan is to do that anyway)